feat: Use glow for Blaze Solver
This replaces the previous method of setting the shader color of the Blaze model.

A potential issue that may occur from this is that glow is visible through walls. We might not want the player to have the glow on their screen while they are not physically present in the puzzle room.